[C-IRIS] move all remaining cspace_free_polytope.cc out of dev.
